Refactor MSAL instance management and imports
- Moved MSAL instance creation to a new file `src/auth/msalClient.ts` for better organization and modularity.
- Updated imports across the application to reference the new MSAL instance location, ensuring consistent access to the authentication client.
- Removed unnecessary console logs in the API and axios setup files to clean up the codebase.
