+++ src/utils/api.ts
@@ -1,7 +1,7 @@
 import { InteractionRequiredAuthError } from '@azure/msal-browser';
 import axios from 'axios';
 import { interactiveLoginRequest, isInTeams, loginRequest } from '../app/msalConfig';
-import { msalInstance } from '../main';
+import { msalInstance } from '@/auth/msalClient';
 
 function getBaseUrl(): string {
   return import.meta.env.VITE_CHAT_API_URI || '';
@@ -30,21 +30,19 @@ API.interceptors.request.use(async (config) => {
 
     return config;
   } catch (error) {
-    console.error('[MSAL] Token acquisition failed:', error);
+    // console.error('Token acquisition failed:', error);
     
     // If the error is due to interaction required (like consent_required),
     // trigger an interactive authentication flow
     if (error instanceof InteractionRequiredAuthError) {
       const inTeamsEnvironment = isInTeams();
-      console.log('[MSAL] Interactive authentication required, triggering login...');
-      console.log('[MSAL] In Teams environment:', inTeamsEnvironment);
+      // console.log('Interactive auth required. Teams:', inTeamsEnvironment);
       
       try {
         const account = msalInstance.getActiveAccount();
         
         // Use popup in Teams, redirect in web browsers
         if (inTeamsEnvironment) {
-          console.log('[MSAL] In Teams - using popup authentication');
           // Use a very explicit popup configuration for Teams
           await msalInstance.loginPopup({
             ...interactiveLoginRequest,
@@ -52,7 +50,6 @@ API.interceptors.request.use(async (config) => {
             prompt: 'consent', // Explicitly request consent
           });
         } else {
-          console.log('[MSAL] In web browser - using redirect authentication');
           await msalInstance.loginRedirect(interactiveLoginRequest);
           // Note: After redirect, this code won't continue executing
           // The page will reload and the user will be authenticated
@@ -72,16 +69,13 @@ API.interceptors.request.use(async (config) => {
         
         config.headers?.set('Authorization', `Bearer ${tokenResponse.accessToken}`);
       } catch (interactiveError) {
-        console.error('[MSAL] Interactive authentication failed:', interactiveError);
+        // console.error('Interactive authentication failed:', interactiveError);
         
         // If we're in Teams and authentication fails, provide helpful messaging
         if (inTeamsEnvironment) {
-          console.log('[MSAL] Teams authentication failed. This might be due to popup blockers or Teams restrictions.');
-          console.log('[MSAL] User may need to manually refresh the Teams app or try again.');
           // Don't redirect in Teams - just log the error
         } else {
           // In web browsers, we could redirect to login page
-          console.log('[MSAL] Web authentication failed, could redirect to login');
           // window.location.href = '/login';
         }
       }
